add abnormality index to patch filename
Previously, the mask index was not added to name. Adding now to help with tracking.

There's a slight chance that two different abnormalities could by chance have the same metadata and identical pixel patch (in the event they are overlapping abnormalities); causing the patch to be overwritten. This wouldn't really matter in real life since they would be identical patches with identical metadata, but knowing which mask helps with tracking patch sampling.